+++ beanie/odm/timeseries.py
@@ -22,16 +22,21 @@ class TimeSeriesConfig(BaseModel):
     time_field: str
     meta_field: Optional[str] = None
     granularity: Optional[Granularity] = None
-    expire_after_seconds: Optional[float] = None
+    bucket_max_span_seconds: Optional[int] = None
+    bucket_rounding_second: Optional[int] = None
+    expire_after_seconds: Optional[int] = None
 
     def to_dict(self) -> Dict[str, Any]:
-        res: Dict[str, Any] = {}
-        timeseries = {"timeField": self.time_field}
+        timeseries: Dict[str, Any] = {"timeField": self.time_field}
         if self.meta_field is not None:
             timeseries["metaField"] = self.meta_field
         if self.granularity is not None:
             timeseries["granularity"] = self.granularity
-        res["timeseries"] = timeseries
+        if self.bucket_max_span_seconds is not None:
+            timeseries["bucketMaxSpanSeconds"] = self.bucket_max_span_seconds
+        if self.bucket_rounding_second is not None:
+            timeseries["bucketRoundingSeconds"] = self.bucket_rounding_second
+        res: Dict[str, Any] = {"timeseries": timeseries}
         if self.expire_after_seconds is not None:
             res["expireAfterSeconds"] = self.expire_after_seconds
         return res