+++ tests/odm/query/test_aggregate.py
@@ -6,7 +6,7 @@ from pydantic.main import BaseModel
 from pymongo.errors import OperationFailure
 
 from beanie.odm.fields import SortDirection
-from tests.odm.models import Sample
+from tests.odm.models import DocumentWithTextIndexAndLink, Sample
 
 
 async def test_aggregate(preset_documents):
@@ -140,3 +140,51 @@ async def test_clone(preset_documents):
         {"$group": {"_id": "$string", "total": {"$sum": "$integer"}}},
         {"a": "b"},
     ]
+
+
+@pytest.mark.parametrize("text_query_count", [0, 1, 2])
+@pytest.mark.parametrize("non_text_query_count", [0, 1, 2])
+async def test_with_text_queries(
+    text_query_count: int, non_text_query_count: int
+):
+    text_query = {"$text": {"$search": "text_search"}}
+    non_text_query = {"s": "test_string"}
+    aggregation_pipeline = [{"$count": "count"}]
+    queries = []
+
+    if text_query_count:
+        queries.append(text_query)
+        if text_query_count > 1:
+            queries.append(text_query)
+
+    if non_text_query_count:
+        queries.append(non_text_query)
+        if non_text_query_count > 1:
+            queries.append(non_text_query)
+
+    query = DocumentWithTextIndexAndLink.find(*queries, fetch_links=True)
+
+    expected_aggregation_pipeline = []
+    if text_query_count:
+        expected_aggregation_pipeline.append(
+            {"$match": text_query}
+            if text_query_count == 1
+            else {"$match": {"$and": [text_query, text_query]}}
+        )
+
+    for link_info in query.document_model.get_link_fields().values():
+        expected_aggregation_pipeline.extend(link_info.iter_pipeline_stages())
+
+    if non_text_query_count:
+        expected_aggregation_pipeline.append(
+            {"$match": non_text_query}
+            if non_text_query_count == 1
+            else {"$match": {"$and": [non_text_query, non_text_query]}}
+        )
+
+    expected_aggregation_pipeline.extend(aggregation_pipeline)
+
+    assert (
+        query.build_aggregation_pipeline(*aggregation_pipeline)
+        == expected_aggregation_pipeline
+    )