+++ tests/odm/test_encoder.py
@@ -1,5 +1,6 @@
 import re
 from datetime import date, datetime
+from uuid import uuid4
 
 from bson import Binary, Regex
 from pydantic import AnyUrl
@@ -9,6 +10,7 @@ from tests.odm.models import (
     Child,
     DocumentForEncodingTest,
     DocumentForEncodingTestDate,
+    DocumentWithComplexDictKey,
     DocumentWithDecimalField,
     DocumentWithHttpUrlField,
     DocumentWithKeepNullsFalse,
@@ -149,3 +151,18 @@ async def test_should_be_able_to_save_retrieve_doc_with_url():
 
     assert isinstance(new_doc.url_field, AnyUrl)
     assert new_doc.url_field == doc.url_field
+
+
+async def test_dict_with_complex_key():
+    assert isinstance(Encoder().encode({uuid4(): datetime.now()}), dict)
+
+    uuid = uuid4()
+    # reset microseconds, because it looses by mongo
+    dt = datetime.now().replace(microsecond=0)
+
+    doc = DocumentWithComplexDictKey(dict_field={uuid: dt})
+    await doc.insert()
+    new_doc = await DocumentWithComplexDictKey.get(doc.id)
+
+    assert isinstance(new_doc.dict_field, dict)
+    assert new_doc.dict_field.get(uuid) == dt